refactor: rulebasedbodytype 모듈 core로 이동 및 변수명 매핑 수정
