+++ backend/schemas/body_type.py
@@ -55,6 +55,17 @@ class BodyTypeAnalysisInput(BaseModel):
         Returns:
             BodyTypeAnalysisInput: 체형 분석 입력 모델
         """
+        # 키 매핑 함수
+        def map_seg_keys(seg_data: Dict) -> Dict:
+            mapping = {
+                "왼쪽팔": "왼팔",
+                "오른쪽팔": "오른팔",
+                "복부": "몸통",
+                "왼쪽하체": "왼다리",
+                "오른쪽하체": "오른다리"
+            }
+            return {mapping.get(k, k): v for k, v in seg_data.items()}
+
         return cls(
             성별=inbody.기본정보.성별,
             연령=inbody.기본정보.연령,
@@ -63,8 +74,8 @@ class BodyTypeAnalysisInput(BaseModel):
             BMI=inbody.비만분석.BMI,
             체지방률=inbody.비만분석.체지방률,
             골격근량=inbody.체중관리.골격근량,
-            muscle_seg=MuscleFatSegment(**muscle_seg),
-            fat_seg=MuscleFatSegment(**fat_seg)
+            muscle_seg=MuscleFatSegment(**map_seg_keys(muscle_seg)),
+            fat_seg=MuscleFatSegment(**map_seg_keys(fat_seg))
         )
     
     class Config:
